command_line/src/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile /Users/minhtran/Documents/WORK/training/minhtran/data-learning/kedro/command_line/src/requirements.in
#
anyconfig==0.9.11
    # via kedro
appdirs==1.4.4
    # via black
appnope==0.1.2
    # via
    #   ipykernel
    #   ipython
argon2-cffi==20.1.0
    # via notebook
arrow==0.17.0
    # via jinja2-time
async-generator==1.10
    # via nbclient
attrs==20.3.0
    # via
    #   black
    #   jsonschema
    #   pytest
backcall==0.2.0
    # via ipython
binaryornot==0.4.4
    # via cookiecutter
black==v19.10b0
    # via -r /Users/minhtran/Documents/WORK/training/minhtran/data-learning/kedro/command_line/src/requirements.in
bleach==3.2.1
    # via nbconvert
cachetools==4.2.0
    # via kedro
certifi==2020.12.5
    # via requests
cffi==1.14.4
    # via argon2-cffi
chardet==4.0.0
    # via
    #   binaryornot
    #   requests
click==7.1.2
    # via
    #   black
    #   cookiecutter
    #   kedro
    #   pip-tools
cookiecutter==1.7.2
    # via kedro
coverage==5.3.1
    # via pytest-cov
decorator==4.4.2
    # via ipython
defusedxml==0.6.0
    # via nbconvert
entrypoints==0.3
    # via nbconvert
flake8==3.8.4
    # via -r /Users/minhtran/Documents/WORK/training/minhtran/data-learning/kedro/command_line/src/requirements.in
fsspec==0.8.5
    # via kedro
gitdb==4.0.5
    # via gitpython
gitpython==3.1.12
    # via kedro
idna==2.10
    # via requests
importlib-metadata==3.4.0
    # via
    #   flake8
    #   jsonschema
    #   pluggy
    #   pytest
iniconfig==1.1.1
    # via pytest
ipykernel==5.4.3
    # via
    #   ipywidgets
    #   jupyter
    #   jupyter-console
    #   notebook
    #   qtconsole
ipython-genutils==0.2.0
    # via
    #   jupyterlab
    #   nbformat
    #   notebook
    #   qtconsole
    #   traitlets
ipython==7.10
    # via
    #   -r /Users/minhtran/Documents/WORK/training/minhtran/data-learning/kedro/command_line/src/requirements.in
    #   ipykernel
    #   ipywidgets
    #   jupyter-console
ipywidgets==7.6.3
    # via jupyter
isort==4.3.21
    # via -r /Users/minhtran/Documents/WORK/training/minhtran/data-learning/kedro/command_line/src/requirements.in
jedi==0.18.0
    # via ipython
jinja2-time==0.2.0
    # via cookiecutter
jinja2==2.11.2
    # via
    #   cookiecutter
    #   jinja2-time
    #   nbconvert
    #   notebook
jmespath==0.10.0
    # via kedro
jsonschema==3.2.0
    # via
    #   jupyterlab-launcher
    #   nbformat
jupyter-client==6.1.11
    # via
    #   -r /Users/minhtran/Documents/WORK/training/minhtran/data-learning/kedro/command_line/src/requirements.in
    #   ipykernel
    #   jupyter-console
    #   kedro
    #   nbclient
    #   notebook
    #   qtconsole
jupyter-console==6.2.0
    # via jupyter
jupyter-core==4.7.0
    # via
    #   jupyter-client
    #   nbconvert
    #   nbformat
    #   notebook
    #   qtconsole
jupyter==1.0.0
    # via -r /Users/minhtran/Documents/WORK/training/minhtran/data-learning/kedro/command_line/src/requirements.in
jupyterlab-launcher==0.10.5
    # via jupyterlab
jupyterlab-pygments==0.1.2
    # via nbconvert
jupyterlab-widgets==1.0.0
    # via ipywidgets
jupyterlab==0.31.1
    # via -r /Users/minhtran/Documents/WORK/training/minhtran/data-learning/kedro/command_line/src/requirements.in
kedro==0.17.0
    # via -r /Users/minhtran/Documents/WORK/training/minhtran/data-learning/kedro/command_line/src/requirements.in
markupsafe==1.1.1
    # via
    #   cookiecutter
    #   jinja2
mccabe==0.6.1
    # via flake8
mistune==0.8.4
    # via nbconvert
nbclient==0.5.1
    # via nbconvert
nbconvert==6.0.7
    # via
    #   jupyter
    #   notebook
nbformat==5.0.8
    # via
    #   ipywidgets
    #   nbclient
    #   nbconvert
    #   nbstripout
    #   notebook
nbstripout==0.3.3
    # via -r /Users/minhtran/Documents/WORK/training/minhtran/data-learning/kedro/command_line/src/requirements.in
nest-asyncio==1.4.3
    # via nbclient
notebook==6.1.6
    # via
    #   jupyter
    #   jupyterlab
    #   jupyterlab-launcher
    #   widgetsnbextension
packaging==20.8
    # via
    #   bleach
    #   pytest
pandocfilters==1.4.3
    # via nbconvert
parso==0.8.1
    # via jedi
pathspec==0.8.1
    # via black
pexpect==4.8.0
    # via ipython
pickleshare==0.7.5
    # via ipython
pip-tools==5.5.0
    # via kedro
pluggy==0.13.1
    # via
    #   kedro
    #   pytest
poyo==0.5.0
    # via cookiecutter
prometheus-client==0.9.0
    # via notebook
prompt-toolkit==3.0.10
    # via
    #   ipython
    #   jupyter-console
ptyprocess==0.7.0
    # via
    #   pexpect
    #   terminado
py==1.10.0
    # via pytest
pycodestyle==2.6.0
    # via flake8
pycparser==2.20
    # via cffi
pyflakes==2.2.0
    # via flake8
pygments==2.7.3
    # via
    #   ipython
    #   jupyter-console
    #   jupyterlab-pygments
    #   nbconvert
    #   qtconsole
pyparsing==2.4.7
    # via packaging
pyrsistent==0.17.3
    # via jsonschema
pytest-cov==2.10.1
    # via -r /Users/minhtran/Documents/WORK/training/minhtran/data-learning/kedro/command_line/src/requirements.in
pytest-mock==1.13.0
    # via -r /Users/minhtran/Documents/WORK/training/minhtran/data-learning/kedro/command_line/src/requirements.in
pytest==6.1.2
    # via
    #   -r /Users/minhtran/Documents/WORK/training/minhtran/data-learning/kedro/command_line/src/requirements.in
    #   pytest-cov
    #   pytest-mock
python-dateutil==2.8.1
    # via
    #   arrow
    #   jupyter-client
python-json-logger==0.1.11
    # via kedro
python-slugify==4.0.1
    # via cookiecutter
pyyaml==5.3.1
    # via kedro
pyzmq==20.0.0
    # via
    #   jupyter-client
    #   notebook
    #   qtconsole
qtconsole==5.0.1
    # via jupyter
qtpy==1.9.0
    # via qtconsole
regex==2020.11.13
    # via black
requests==2.25.1
    # via cookiecutter
send2trash==1.5.0
    # via notebook
six==1.15.0
    # via
    #   argon2-cffi
    #   bleach
    #   cookiecutter
    #   jsonschema
    #   python-dateutil
smmap==3.0.4
    # via gitdb
terminado==0.9.2
    # via notebook
testpath==0.4.4
    # via nbconvert
text-unidecode==1.3
    # via python-slugify
toml==0.10.2
    # via
    #   black
    #   kedro
    #   pytest
toposort==1.6
    # via kedro
tornado==6.1
    # via
    #   ipykernel
    #   jupyter-client
    #   notebook
    #   terminado
traitlets==5.0.5
    # via
    #   ipykernel
    #   ipython
    #   ipywidgets
    #   jupyter-client
    #   jupyter-core
    #   nbclient
    #   nbconvert
    #   nbformat
    #   notebook
    #   qtconsole
typed-ast==1.4.2
    # via black
typing-extensions==3.7.4.3
    # via importlib-metadata
urllib3==1.26.2
    # via requests
wcwidth==0.2.5
    # via prompt-toolkit
webencodings==0.5.1
    # via bleach
wheel==0.32.2
    # via -r /Users/minhtran/Documents/WORK/training/minhtran/data-learning/kedro/command_line/src/requirements.in
widgetsnbextension==3.5.1
    # via ipywidgets
zipp==3.4.0
    # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
